requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
gunicorn==20.1.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.4
pytest-django==4.10.0
redis==5.1.0
sqlparse==0.5.3
